requirements.txt
gradio==2.0.9
numpy==1.21.0
Pillow==8.3.1
tensorflow==2.5.0
tensorflow_hub==0.12.0
